# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: SMIDIV API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
# basic authentication definition
securityDefinitions:
  basicAuth:
    type: basic
  tokenAuth:
    type: apiKey
    in: header
    name: X-API-KEY
paths:
...
  /ubicacion:
    x-swagger-router-controller: ubicacion
    get:
      description: Obtener ubicacion
      operationId: getUbicacion      
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: añadir Ubicacion
      operationId: addUbicacion
      parameters:
        - name: ubicacion
          in: body
          required: true
          schema:
            $ref: "#/definitions/Ubicacion"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /ubicacion/{vehiculo}:
    x-swagger-router-controller: ubicacion
    get:
      security:
        - basicAuth: []
        - tokenAuth: []
      description: Gets the user description for a given username
      operationId: getUbicaciones
      parameters:
        - name: vehiculo
          in: path
          description: The unique username of the user to get the data
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /clave:
    x-swagger-router-controller: SMIDIVID
    get:
      security:
        - basicAuth: []
        - tokenAuth: []
      description: gets the marcaList
      operationId: getClave
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      # require basic authentication on this route
      security:
        - basicAuth: []
        - tokenAuth: []
      description: Adds a new marca
      operationId: addClave
      parameters:
        - name: smidivid
          description: New Marca's information
          in: body
          required: true
          schema:
            $ref: "#/definitions/SMIDIVID"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /brand:
    x-swagger-router-controller: marca
    get:
      security:
        - basicAuth: []
        - tokenAuth: []
      description: gets the marcaList
      operationId: getMarcaList
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      # require basic authentication on this route
      security:
        - basicAuth: []
        - tokenAuth: []
      description: Adds a new marca
      operationId: addMarca
      parameters:
        - name: marca
          description: New Marca's information
          in: body
          required: true
          schema:
            $ref: "#/definitions/Marca"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /user:
    # our controller name
    x-swagger-router-controller: user
    get:
      # require basic authentication on this route
      security:
        - basicAuth: []
        - tokenAuth: []
      description: Gets the users list
      operationId: getUsersList
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUsersListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Adds a new user
      operationId: addUser
      parameters:
        - name: user
          description: New user's information
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AuthenticationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/{username}:
    # our controller name
    x-swagger-router-controller: user
    patch:
      security:
        - basicAuth: []
        - tokenAuth: []
      description: Updates a user by username
      operationId: updateUser
      parameters:
        - in: path
          name: username
          type: string
          required: true
          description: Username of the user to update
        - name: updates
          description: User updates
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserUpdates"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/StringResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      security:
        - basicAuth: []
        - tokenAuth: []
      description: Gets the user description for a given username
      operationId: getUser
      parameters:
        - name: username
          in: path
          description: The unique username of the user to get the data
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      security:
        - basicAuth: []
        - tokenAuth: []
      description: Deletes the user for the given username
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: The unique username of the user to delete
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/StringResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  AuthenticationResponse:
    required:
      - success
      - message
      - token
    properties:
      success:
        type: boolean
      message:
        type: string
      token:
        type: string
  GetUsersListResponse:
    required:
      - users
    properties:
      # The array of users
      users:
        type: array
        items:
          type: object
          properties:
            username:
              type: string
  Marca:
    type: object
    properties:
      nombre: 
        type: string
...
  SMIDIVID:
    type: object
    properties:
      clave: 
        type: string
        description: Clave de identificacion SMIDIV
    required:
      - clave
...
  Ubicacion:
    type: object
    properties:
      smidivID:
        type: string
        description: Ubicacion ligada al automovil
      ubicacion:
        type: object
        properties:
          lat:
            type: number
          lng: 
            type: number
      fechaCreacion:
        type: string
    required:
    - smidivID
...
  User:
    type: object
    properties:
      username:
        type: string
        description: user's unique username
      email:
        type: string
        description: user's email
      password:
        type: string
        format: password
      records:
        type: array
        items:
          type: object
    required:
      - username
      - email
      - password
  UserUpdates:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
        format: password
      email:
        type: string
      profile:
        type: object
  StringResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  GetUserResponse:
    properties:
      user:
        type: object
        properties:
          _id:
            type: string
          username:
            type: string
          email:
            type: string
          profile:
            type: object
          routines:
            type: array
            items:
              type: object
          records:
            type: array
            items:
              type: object
    required:
      - user
  OkResponse:
    required:
      - response
      - success
    properties:
      response:
        type: object
        properties: {}
      success:
        type: boolean
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string